Refactor CSV output
